refactor: unify broadcastTxs logic and remove process_broadcast_multiple
